requirements.txt
langchain
langchain-teddynote
fastapi
uvicorn
python-dotenv
psycopg2
PyPDF2
chromadb
tiktoken